moves cnl/bits/fixed_point_*.h into cnl/bits/fixed_point/*.h
