+++ include/cnl/bits/fixed_point/math.h
