+++ include/cnl/num_traits.h
@@ -11,6 +11,7 @@
 #define CNL_NUMERIC_TRAITS 1
 
 #include "limits.h"
+
 #include "bits/type_traits.h"
 
 #include <utility>
