+++ src/test/fixed_point/constants.cpp
