+++ src/test/fixed_point/extras.cpp
